test(compiler-plugin/previewCollection/fir): split FIR tests by logic (hintGeneration / hintAndMarkerGeneration / scopeValidation)
Ticket 1 で確定した production の FIR logic 構造に sync する。本 commit では FIR 側の
plugin 挙動を検証するテスト群を logic 別 (hintGeneration / hintAndMarkerGeneration /
scopeValidation) に再配置する。

`fir/hintGeneration/` (hint stub 関数 = `previewHint_<scope>` の挙動):

- `BacktickIdentifierSanitizationTest.kt` (識別子サニタイズ)
- `PreviewHintEmissionTest.kt` (hint 関数の emission)
- `PreviewHintHiddenDeprecationTest.kt` (`@Deprecated(HIDDEN)` 関連)
- `PreviewHintScopeIgnoreInteropTest.kt` (scope と `ignored` の interop; PR #187)

`fir/hintAndMarkerGeneration/` (production で hint stub + marker interface が
`PreviewHintFirGenerator` に統合されたため、marker 単体検証もここに置く):

- `PreviewHintMarkerSealOrHiddenTest.kt` (marker の sealed / hidden 検証)

`fir/scopeValidation/` (`@ComposePreviewLabOption(collectScopes)` /
`collect[All]ModulePreviews(scope = ...)` の literal 検証):

- `PreviewHintScopeValidationTest.kt` (PR #187)
- `DefaultCollectScopeDslTest.kt` (`DefaultCollectScope` sentinel + 既定 routing)

package: `feature` → `feature.previewCollection.fir.<logic>`

`PreviewHintHiddenDeprecationTest` / `PreviewHintMarkerSealOrHiddenTest` /
`DefaultCollectScopeDslTest` は `PreviewHintTestSupport` の internal extension
(`previewHint*Names()`) に依存するため対応する `import me.tbsten.compose.preview.lab.compiler.feature.previewHint*`
を追加。

設計逸脱メモ:
Ticket 仕様では `fir/markerGeneration/` を新設する想定だったが、Ticket 1
landing 時点の production は `PreviewHintFirGenerator` を分割せず hint stub +
marker interface を同 generator にまとめた `fir/hintAndMarkerGeneration/`
となっていた。production 構造への sync が本チケットの目的なので production
に追従し、`fir/hintAndMarkerGeneration/` を採用した。

検証:
- `:compiler-plugin:ktlintFormat` + `:compiler-plugin:ktlintCheck` pass
- `:compiler-plugin:test --rerun-tasks` pass (131 tests total)
- `apiCheck` pass
- `find compiler-plugin -name build/api -type d` で `compat-k*/build/api/` 以外ゼロ
- wildcard import 増加なし
